A measurement is kept at the width it was measured.
Three lossy transforms sat between the detector and the row, each one
deliberate and none of them paid for:

  landmark_2d_106   `round(x, 5)` on the normalized coordinate
  landmark_3d_68    the same, plus `round(z, 2)` on the depth
  pose              `round(deg, 2)` on all three angles
  landmarks         a float32 blob holding NORMALIZED coordinates, which a
                    read multiplies back up by the frame size

Measured, not argued. `compat/consumers/gallery_storage.py` replays the
producer's own emission against this application's write path at rtol=0,
atol=0, and found 19 keys that do not survive: `kps` to 2.4e-4 px,
`landmark_2d_106` to 0.0322 px, `landmark_3d_68` to 0.0320, `pose` to 0.0034
degrees, on every face this application has ever stored. float32's half-ulp
in [0, 1] is 2**-24, which on a 6528 px frame is 3.9e-4 source pixels
discarded before anything asks for the value.

vision/faces.py stops rounding, db/detect.py writes float64, and connect.py
and schema.sql go to v46. Step 45 widens every stored blob: float32
normalized coordinates widen to float64 exactly, so the rows read back at the
precision the next detection writes at. It converts unconditionally --
`user_version` is what says these bytes are float32, and width discriminates
nothing, since five float32 pairs and five float64 pairs are 40 and 80 bytes
and both divide by 8.

What v46 CANNOT do: the rounded columns are gone. `round(x, 5)` destroyed the
digits and there is nothing in the row to recover them from. Those keep what
they have until a faces pass reads the original bytes again
(`POST /jobs/faces` with `{"everything": true}`), and this time it writes what
the detector produced.

The same lane also loses its ablation. Its retained state holds exactly one
key, so removing that key and reading it back could only ever report "broke"
-- for every key, including one nothing needs. That vacuous experiment was
the SOLE necessity evidence behind pose, age, gender, bbox, det_score, kps
and landmark_2d_106 in answer.json. Conservation is what this lane measures;
necessity belongs to the lanes that replay a real consumer boundary.

Three new compat modules, so "minimum canonical evidence" means what it says:

  storage/precision.py    a width substitution per retained numeric value --
                          float16 round trip, original dtype restored -- so
                          minimum is bytes and not only keys. float16 because
                          it is a storable width with an IEEE 754 error, not
                          a constant chosen here.
  storage/derivatives.py  the two candidate answers a whole-reference
                          consumer has, both built by executing the
                          application's own code: `preview`, what the store
                          returns today at 1440 and lossy, and `lossless`,
                          what it would have to add. PNG, because a store
                          holds bytes and answer.json must say what the
                          minimum state costs.
  corpus/cache.py         the memo, persisted across the six interpreters
                          sharded.py splits the population into. Namespaced
                          on identity()["digest"], so a stale entry is never
                          found rather than found and trusted, and every
                          entry is read back and compared before it is kept.

Dead code with it: Emission.bytes_total, Observation.has,
Observation.bytes_total, groups.summarise, GalleryStorageRunner.ablate.

`just check` green at 8.06s.

+++ db/detect.py
@@ -103,7 +103,14 @@ def harvest(
         if found.landmarks:
             import numpy as np
 
-            record["landmarks"] = np.asarray(found.landmarks, dtype=np.float32).tobytes()
+            # float64, not float32. The blob holds NORMALIZED coordinates,
+            # so reading one back multiplies by the frame size -- and a
+            # float32 normalized value carries a half-ulp of 6e-8, which on a
+            # 6528 px frame is 3.9e-4 source pixels the producer measured and
+            # the store then threw away. Measured at up to 2.4e-4 px by
+            # `compat/consumers/gallery_storage.py`. Ten doubles per face is
+            # 80 bytes against 40; the precision is worth more than 40 bytes.
+            record["landmarks"] = np.asarray(found.landmarks, dtype=np.float64).tobytes()
         # EVERYTHING the backend said, then promotions out of it. Not an
         # allowlist: this copied `age` and `sex` and dropped pose and two
         # dense landmark sets on the line that read them, so three columns
